// MegabinFE/src/app/services/model/addressSuggestion.ts
/**
 * MySmartCloset API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Location } from './location';

export interface AddressSuggestion {
  label?: string | null;
  location?: Location;
  region?: string | null;
  locality?: string | null;
  country?: string | null;
}